:truck: Move VersionArgs and UndoArgs from main.rs into their own command modules
