+++ src/command/version.rs
@@ -1,4 +1,3 @@
-use crate::VersionArgs;
 use cargo_semantic_release::{
     create_release_tag, render_tag, Changes, RepositoryExtension, SemanticReleaseConfig,
     SemanticVersionAction,
@@ -7,6 +6,48 @@ use git2::{Oid, Repository};
 use semver::Version;
 use std::path::Path;
 use std::{env, process};
+
+#[derive(clap::Args)]
+pub struct VersionArgs {
+    /// Print the next version's tag (e.g. `v1.2.3`) instead of the bare version
+    #[arg(long)]
+    print_tag: bool,
+
+    /// Force a major version bump instead of deriving it from commit history
+    #[arg(long, conflicts_with_all = ["minor", "patch"])]
+    major: bool,
+
+    /// Force a minor version bump instead of deriving it from commit history
+    #[arg(long, conflicts_with = "patch")]
+    minor: bool,
+
+    /// Force a patch version bump instead of deriving it from commit history
+    #[arg(long)]
+    patch: bool,
+
+    /// Skip creating a commit for the version bump
+    #[arg(long)]
+    no_commit: bool,
+
+    /// Skip pushing the version-bump commit and any created tags to origin
+    #[arg(long)]
+    no_push: bool,
+}
+
+impl VersionArgs {
+    /// The [`SemanticVersionAction`] forced by `--major`/`--minor`/`--patch`, if any.
+    fn forced_action(&self) -> Option<SemanticVersionAction> {
+        if self.major {
+            Some(SemanticVersionAction::IncrementMajor)
+        } else if self.minor {
+            Some(SemanticVersionAction::IncrementMinor)
+        } else if self.patch {
+            Some(SemanticVersionAction::IncrementPatch)
+        } else {
+            None
+        }
+    }
+}
 
 pub fn run_version_command(args: VersionArgs, verbosity: u8, noop: bool) {
     let path = env::current_dir().unwrap_or_else(|error| {
